# static/openapi.yaml
openapi: 3.0.3
info:
  title: Api-SportMatch
  description: >
    API REST de SportMatch. Fase 2: autenticación, usuarios y administradores.
    Fase 3 (en construcción): deportes, cuestionario de registro, eventos,
    inscripciones y valoraciones, organizadores y validación, administración
    y soporte. Ver /docs/entidades_completas.md en el repositorio para el
    catálogo completo de las 89 entidades del modelo.
  version: 0.5.0
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/reportes:
    post:
      summary: Reportar contenido o a un usuario (tipo_entidad/entidad_id son genéricos)
      tags: [Administración y soporte]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tipo_entidad, entidad_id, motivo]
              properties:
                tipo_entidad: { type: string, example: "evento" }
                entidad_id: { type: integer }
                motivo: { type: string }
                descripcion: { type: string }
      responses:
        "201": { description: Reporte creado }
    get:
      summary: Bandeja de moderación (solo admin; ?estado=pendiente por defecto)
      tags: [Administración y soporte]
      responses:
        "200": { description: Lista de reportes }
  /api/reportes/{reporte_id}:
    put:
      summary: Moderar un reporte (solo admin) — se registra en la bitácora de auditoría
      tags: [Administración y soporte]
      parameters:
        - { name: reporte_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                estado: { type: string, enum: [pendiente, en_revision, resuelto, desestimado] }
                accion_tomada: { type: string }
                notas_moderador: { type: string }
      responses:
        "200": { description: Reporte actualizado }
  /api/tickets:
    post:
      summary: Abrir un ticket de soporte con el mensaje inicial
      tags: [Administración y soporte]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [asunto, mensaje]
              properties:
                asunto: { type: string }
                mensaje: { type: string }
                categoria: { type: string }
                prioridad: { type: string, enum: [baja, media, alta, urgente] }
      responses:
        "201": { description: Ticket creado }
    get:
      summary: Bandeja de soporte (solo admin; filtros estado, prioridad, asignado_a)
      tags: [Administración y soporte]
      responses:
        "200": { description: Lista de tickets }
  /api/tickets/{ticket_id}:
    get:
      summary: Detalle de un ticket con sus mensajes (dueño no ve mensajes internos)
      tags: [Administración y soporte]
      parameters:
        - { name: ticket_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Ticket con mensajes }
    put:
      summary: Asignar, priorizar o cerrar un ticket (solo admin)
      tags: [Administración y soporte]
      parameters:
        - { name: ticket_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Ticket actualizado }
  /api/tickets/{ticket_id}/mensajes:
    post:
      summary: Agregar un mensaje al ticket (dueño o admin; solo admin puede marcarlo es_interno)
      tags: [Administración y soporte]
      parameters:
        - { name: ticket_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mensaje]
              properties:
                mensaje: { type: string }
                es_interno: { type: boolean }
      responses:
        "201": { description: Mensaje agregado }
  /api/admin/auditoria:
    get:
      summary: Bitácora de auditoría de solo lectura (solo admin)
      tags: [Administración y soporte]
      parameters:
        - { name: tipo_entidad, in: query, schema: { type: string } }
        - { name: entidad_id, in: query, schema: { type: integer } }
        - { name: usuario_id, in: query, schema: { type: integer } }
      responses:
        "200": { description: Lista paginada de registros de auditoría }